requirement.txt
async-timeout==4.0.2
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
cycler==0.10.0
Flask==1.1.2
html5lib==1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
kafka-python==2.0.2
kiwisolver==1.2.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.18.2
pandas==1.0.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==1.0.0
pytz==2019.3
redis==4.6.0
requests==2.23.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
